Huge improvements to PreProcessMousePosition_auto
Automates mouse tracking for places where the function was previously
flagging from auto_thresh.

INTUITION:
Where the mouse is moving impossibly fast or is at (0,0), take the first
frame of the video and subtract from the frame where Cineplex lost
track.
Do some fancy smoothing and image processing. Look for regions that
could possibly be the size of a mouse.
If more than one possible blob could be the mouse, look for previous
location of mouse and pick the blob that's closest.
If that fails, user input.

Finally, this automation is not perfect. If there are any points you're
not satisfied with, enter "m" when normally prompted for a flaw that
needs to be corrected to correct the old fashioned way with mouse
clicks.